refactor(workspaces): worktree 命令簇并入 worktree.rs
- commands.rs 2604→2224（两段累计 -1269）；worktree.rs 32→415 行
- spawn_with_app/cleanup_engine_sessions_for_workspace 升 pub(crate) 供兄弟模块
- mod.rs 补 worktree glob 重导出保 command_registry 零改动
- baseline 重生成 + 新文件登记豁免
验证：cargo check 零 error；workspaces 114 passed；rustfmt 过；gate 绿
