+++ docs/architecture/large-file-baseline.json
@@ -1,5 +1,5 @@
 {
-  "generatedAt": "2026-08-30T18:44:02.052Z",
+  "generatedAt": "2026-08-30T18:58:57.017Z",
   "scope": "fail",
   "policyVersion": "2026-08-30.policy-v5",
   "entries": [
@@ -291,16 +291,6 @@
       "priority": "P2",
       "warnThreshold": 2600,
       "failThreshold": 3000,
-      "thresholdSource": "policy"
-    },
-    {
-      "path": "src/features/composer/components/Composer.tsx",
-      "lines": 3085,
-      "type": "ts/tsx",
-      "policyId": "feature-hotpath",
-      "priority": "P1",
-      "warnThreshold": 2400,
-      "failThreshold": 2800,
       "thresholdSource": "policy"
     },
     {
