+++ docs/architecture/large-file-new-file-baseline.md
@@ -8,7 +8,7 @@ status: generated
 > **Lifecycle**: Generated snapshot. This report reflects repository state only at the generation time below; rerun the current scanner before making governance decisions.
 > **Fact boundary**: This Markdown is a human-readable projection, not an accepted baseline source or proof that the current tree passes the gate. Machine-readable baseline JSON and current CLI output remain authoritative.
 
-- Generated at: 2026-08-30T18:44:03.160Z
+- Generated at: 2026-08-30T18:58:57.922Z
 - Scope: new-file
 - Policy version: 2026-08-30.policy-v5
 - Count: 395
@@ -44,7 +44,6 @@ status: generated
 | `src/features/files/components/FileViewPanel.tsx` | 3149 | ts/tsx | default-source | P1 | 2600 | 800 | fail | captured |  |  |
 | `src/features/git-history/components/GitHistoryPanel.test.tsx` | 3134 | ts/tsx | test-files | P2 | 2600 | 800 | fail | captured |  |  |
 | `src/features/threads/loaders/historyLoaders.test.ts` | 3089 | ts/tsx | test-files | P2 | 2600 | 800 | fail | captured |  |  |
-| `src/features/composer/components/Composer.tsx` | 3085 | ts/tsx | feature-hotpath | P1 | 2400 | 800 | fail | captured |  |  |
 | `src/features/threads/hooks/useThreadEventHandlers.test.ts` | 3078 | ts/tsx | test-files | P2 | 2600 | 800 | fail | captured |  |  |
 | `src/styles/diff.css` | 3077 | css | styles | P1 | 2200 | 800 | fail | captured |  |  |
 | `src/features/threads/hooks/useThreadActions.ts` | 3066 | ts/tsx | feature-hotpath | P1 | 2400 | 800 | fail | captured |  |  |
@@ -69,9 +68,9 @@ status: generated
 | `src/features/files/components/FileTreePanel.tsx` | 2622 | ts/tsx | default-source | P1 | 2600 | 800 | fail | captured |  |  |
 | `src-tauri/src/engine/grok_history.rs` | 2620 | rust | bridge-runtime-critical | P0 | 2200 | 800 | fail | captured |  |  |
 | `src/features/settings/components/SettingsView.tsx` | 2611 | ts/tsx | feature-hotpath | P1 | 2400 | 800 | fail | captured |  |  |
-| `src-tauri/src/workspaces/commands.rs` | 2606 | rust | default-source | P1 | 2600 | 800 | fail | captured |  |  |
 | `src-tauri/src/runtime/mod.rs` | 2585 | rust | bridge-runtime-critical | P0 | 2200 | 800 | fail | captured |  |  |
 | `src-tauri/src/local_usage/tests.rs` | 2577 | rust | default-source | P1 | 2600 | 800 | fail | captured |  |  |
+| `src/features/composer/components/Composer.tsx` | 2573 | ts/tsx | feature-hotpath | P1 | 2400 | 800 | fail | captured |  |  |
 | `src-tauri/src/bin/cc_gui_daemon.rs` | 2562 | rust | default-source | P1 | 2600 | 800 | fail | captured |  |  |
 | `src-tauri/src/engine/commands_send.rs` | 2553 | rust | bridge-runtime-critical | P0 | 2200 | 800 | fail | captured |  |  |
 | `src-tauri/src/email/session_continuation.rs` | 2551 | rust | default-source | P1 | 2600 | 800 | fail | captured |  |  |
@@ -97,6 +96,7 @@ status: generated
 | `src-tauri/src/engine/claude/tests_core.rs` | 2264 | rust | bridge-runtime-critical | P0 | 2200 | 800 | fail | captured |  |  |
 | `src/styles/tool-blocks.css` | 2249 | css | styles | P1 | 2200 | 800 | fail | captured |  |  |
 | `src/features/engine/hooks/useEngineController.test.tsx` | 2230 | ts/tsx | test-files | P2 | 2600 | 800 | fail | captured |  |  |
+| `src-tauri/src/workspaces/commands.rs` | 2224 | rust | default-source | P1 | 2600 | 800 | fail | captured |  |  |
 | `src/styles/session-activity.css` | 2212 | css | styles | P1 | 2200 | 800 | fail | captured |  |  |
 | `src/features/threads/loaders/claudeHistoryLoader.test.ts` | 2202 | ts/tsx | test-files | P2 | 2600 | 800 | fail | captured |  |  |
 | `src-tauri/src/coordinator_tests.rs` | 2195 | rust | default-source | P1 | 2600 | 800 | fail | captured |  |  |
