+++ src-tauri/src/workspaces/worktree.rs
@@ -1,6 +1,25 @@
 use std::path::PathBuf;
 
+use serde_json::json;
+use tauri::{AppHandle, Manager, State};
+use tokio::io::AsyncWriteExt;
+
+use crate::remote_backend;
 use crate::shared::worktree_core;
+use crate::shared::workspaces_core;
+use crate::state::AppState;
+use crate::types::WorkspaceInfo;
+use std::process::Stdio;
+
+use crate::git_utils::resolve_git_root;
+use crate::types::WorktreeSetupStatus;
+use crate::utils::{git_env_path, resolve_git_binary};
+use super::commands::{cleanup_engine_sessions_for_workspace, spawn_with_app};
+use super::git::{
+    git_branch_exists, git_find_remote_for_branch, git_remote_branch_exists, git_remote_exists,
+    is_missing_worktree_error, run_git_command_bytes, run_git_command_owned, run_git_diff,
+    unique_branch_name,
+};
 
 pub(crate) fn sanitize_worktree_name(branch: &str) -> String {
     worktree_core::sanitize_worktree_name(branch)
@@ -29,4 +48,387 @@ pub(crate) fn build_clone_destination_path(copies_folder: &PathBuf, copy_name: &
 
 pub(crate) fn null_device_path() -> &'static str {
     worktree_core::null_device_path()
+}
+
+#[tauri::command]
+pub(crate) async fn add_worktree(
+    parent_id: String,
+    branch: String,
+    base_ref: Option<String>,
+    publish_to_origin: Option<bool>,
+    state: State<'_, AppState>,
+    app: AppHandle,
+) -> Result<WorkspaceInfo, String> {
+    let publish_to_origin = publish_to_origin.unwrap_or(true);
+    if remote_backend::is_remote_mode(&*state).await {
+        let response = remote_backend::call_remote(
+            &*state,
+            app,
+            "add_worktree",
+            json!({
+                "parentId": parent_id,
+                "branch": branch,
+                "baseRef": base_ref,
+                "publishToOrigin": publish_to_origin
+            }),
+        )
+        .await?;
+        return serde_json::from_value(response).map_err(|err| err.to_string());
+    }
+
+    let data_dir = app
+        .path()
+        .app_data_dir()
+        .map_err(|err| format!("Failed to resolve app data dir: {err}"))?;
+
+    workspaces_core::add_worktree_core(
+        parent_id,
+        branch,
+        base_ref,
+        publish_to_origin,
+        &data_dir,
+        &state.workspaces,
+        &state.sessions,
+        &state.app_settings,
+        &state.storage_path,
+        |value| sanitize_worktree_name(value),
+        |root, name| Ok(unique_worktree_path(root, name)),
+        |root, branch| {
+            let root = root.clone();
+            let branch = branch.to_string();
+            async move { git_branch_exists(&root, &branch).await }
+        },
+        None::<fn(&PathBuf, &str) -> std::future::Ready<Result<Option<String>, String>>>,
+        |root, args| {
+            workspaces_core::run_git_command_unit(root, args, |repo, args_owned| {
+                run_git_command_owned(repo, args_owned)
+            })
+        },
+        |entry, default_bin, codex_args, codex_home| {
+            spawn_with_app(&app, entry, default_bin, codex_args, codex_home)
+        },
+    )
+    .await
+}
+
+#[tauri::command]
+pub(crate) async fn worktree_setup_status(
+    workspace_id: String,
+    state: State<'_, AppState>,
+    app: AppHandle,
+) -> Result<WorktreeSetupStatus, String> {
+    if remote_backend::is_remote_mode(&*state).await {
+        let response = remote_backend::call_remote(
+            &*state,
+            app,
+            "worktree_setup_status",
+            json!({ "workspaceId": workspace_id }),
+        )
+        .await?;
+        return serde_json::from_value(response).map_err(|err| err.to_string());
+    }
+
+    let data_dir = app
+        .path()
+        .app_data_dir()
+        .map_err(|err| format!("Failed to resolve app data dir: {err}"))?;
+    workspaces_core::worktree_setup_status_core(&state.workspaces, &workspace_id, &data_dir).await
+}
+
+#[tauri::command]
+pub(crate) async fn worktree_setup_mark_ran(
+    workspace_id: String,
+    state: State<'_, AppState>,
+    app: AppHandle,
+) -> Result<(), String> {
+    if remote_backend::is_remote_mode(&*state).await {
+        remote_backend::call_remote(
+            &*state,
+            app,
+            "worktree_setup_mark_ran",
+            json!({ "workspaceId": workspace_id }),
+        )
+        .await?;
+        return Ok(());
+    }
+
+    let data_dir = app
+        .path()
+        .app_data_dir()
+        .map_err(|err| format!("Failed to resolve app data dir: {err}"))?;
+    workspaces_core::worktree_setup_mark_ran_core(&state.workspaces, &workspace_id, &data_dir).await
+}
+
+#[tauri::command]
+
+pub(crate) async fn remove_worktree(
+    id: String,
+    state: State<'_, AppState>,
+    app: AppHandle,
+) -> Result<(), String> {
+    if remote_backend::is_remote_mode(&*state).await {
+        remote_backend::call_remote(&*state, app, "remove_worktree", json!({ "id": id })).await?;
+        return Ok(());
+    }
+
+    workspaces_core::remove_worktree_core(
+        id.clone(),
+        &state.workspaces,
+        &state.sessions,
+        &state.storage_path,
+        |root, args| {
+            workspaces_core::run_git_command_unit(root, args, |repo, args_owned| {
+                run_git_command_owned(repo, args_owned)
+            })
+        },
+        |error| is_missing_worktree_error(error),
+        |path| {
+            std::fs::remove_dir_all(path)
+                .map_err(|err| format!("Failed to remove worktree folder: {err}"))
+        },
+    )
+    .await?;
+
+    cleanup_engine_sessions_for_workspace(&state, &id)
+        .await
+        .map_err(|error| format!("worktree removed but engine cleanup failed: {error}"))?;
+
+    Ok(())
+}
+
+#[tauri::command]
+pub(crate) async fn rename_worktree(
+    id: String,
+    branch: String,
+    state: State<'_, AppState>,
+    app: AppHandle,
+) -> Result<WorkspaceInfo, String> {
+    if remote_backend::is_remote_mode(&*state).await {
+        let response = remote_backend::call_remote(
+            &*state,
+            app,
+            "rename_worktree",
+            json!({ "id": id, "branch": branch }),
+        )
+        .await?;
+        return serde_json::from_value(response).map_err(|err| err.to_string());
+    }
+
+    let data_dir = app
+        .path()
+        .app_data_dir()
+        .map_err(|err| format!("Failed to resolve app data dir: {err}"))?;
+
+    workspaces_core::rename_worktree_core(
+        id,
+        branch,
+        &data_dir,
+        &state.workspaces,
+        &state.sessions,
+        &state.app_settings,
+        &state.storage_path,
+        |entry| resolve_git_root(entry),
+        |root, name| {
+            let root = root.clone();
+            let name = name.to_string();
+            async move {
+                unique_branch_name(&root, &name, None)
+                    .await
+                    .map(|(branch, _was_suffixed)| branch)
+            }
+        },
+        |value| sanitize_worktree_name(value),
+        |root, name, current| unique_worktree_path_for_rename(root, name, current),
+        |root, args| {
+            workspaces_core::run_git_command_unit(root, args, |repo, args_owned| {
+                run_git_command_owned(repo, args_owned)
+            })
+        },
+        |entry, default_bin, codex_args, codex_home| {
+            spawn_with_app(&app, entry, default_bin, codex_args, codex_home)
+        },
+    )
+    .await
+}
+
+#[tauri::command]
+pub(crate) async fn rename_worktree_upstream(
+    id: String,
+    old_branch: String,
+    new_branch: String,
+    state: State<'_, AppState>,
+    app: AppHandle,
+) -> Result<(), String> {
+    if remote_backend::is_remote_mode(&*state).await {
+        remote_backend::call_remote(
+            &*state,
+            app,
+            "rename_worktree_upstream",
+            json!({ "id": id, "oldBranch": old_branch, "newBranch": new_branch }),
+        )
+        .await?;
+        return Ok(());
+    }
+
+    workspaces_core::rename_worktree_upstream_core(
+        id,
+        old_branch,
+        new_branch,
+        &state.workspaces,
+        |entry| resolve_git_root(entry),
+        |root, branch| {
+            let root = root.clone();
+            let branch = branch.to_string();
+            async move { git_branch_exists(&root, &branch).await }
+        },
+        |root, branch| {
+            let root = root.clone();
+            let branch = branch.to_string();
+            async move { git_find_remote_for_branch(&root, &branch).await }
+        },
+        |root, remote| {
+            let root = root.clone();
+            let remote = remote.to_string();
+            async move { git_remote_exists(&root, &remote).await }
+        },
+        |root, remote, branch| {
+            let root = root.clone();
+            let remote = remote.to_string();
+            let branch = branch.to_string();
+            async move { git_remote_branch_exists(&root, &remote, &branch).await }
+        },
+        |root, args| {
+            workspaces_core::run_git_command_unit(root, args, |repo, args_owned| {
+                run_git_command_owned(repo, args_owned)
+            })
+        },
+    )
+    .await
+}
+
+#[tauri::command]
+pub(crate) async fn apply_worktree_changes(
+    workspace_id: String,
+    state: State<'_, AppState>,
+) -> Result<(), String> {
+    let (entry, parent) = {
+        let workspaces = state.workspaces.lock().await;
+        let entry = workspaces
+            .get(&workspace_id)
+            .cloned()
+            .ok_or("workspace not found")?;
+        if !entry.kind.is_worktree() {
+            return Err("Not a worktree workspace.".to_string());
+        }
+        let parent_id = entry.parent_id.clone().ok_or("worktree parent not found")?;
+        let parent = workspaces
+            .get(&parent_id)
+            .cloned()
+            .ok_or("worktree parent not found")?;
+        (entry, parent)
+    };
+
+    let worktree_root = resolve_git_root(&entry)?;
+    let parent_root = resolve_git_root(&parent)?;
+
+    let parent_status = run_git_command_bytes(&parent_root, &["status", "--porcelain"]).await?;
+    if !String::from_utf8_lossy(&parent_status).trim().is_empty() {
+        return Err(
+            "Your current branch has uncommitted changes. Please commit, stash, or discard them before applying worktree changes."
+                .to_string(),
+        );
+    }
+
+    let mut patch: Vec<u8> = Vec::new();
+    let staged_patch = run_git_diff(
+        &worktree_root,
+        &["diff", "--binary", "--no-color", "--cached"],
+    )
+    .await?;
+    patch.extend_from_slice(&staged_patch);
+    let unstaged_patch = run_git_diff(&worktree_root, &["diff", "--binary", "--no-color"]).await?;
+    patch.extend_from_slice(&unstaged_patch);
+
+    let untracked_output = run_git_command_bytes(
+        &worktree_root,
+        &["ls-files", "--others", "--exclude-standard", "-z"],
+    )
+    .await?;
+    for raw_path in untracked_output.split(|byte| *byte == 0) {
+        if raw_path.is_empty() {
+            continue;
+        }
+        let path = String::from_utf8_lossy(raw_path).to_string();
+        let diff = run_git_diff(
+            &worktree_root,
+            &[
+                "diff",
+                "--binary",
+                "--no-color",
+                "--no-index",
+                "--",
+                null_device_path(),
+                &path,
+            ],
+        )
+        .await?;
+        patch.extend_from_slice(&diff);
+    }
+
+    if String::from_utf8_lossy(&patch).trim().is_empty() {
+        return Err("No changes to apply.".to_string());
+    }
+
+    let git_bin = resolve_git_binary().map_err(|e| format!("Failed to run git: {e}"))?;
+    let mut child = crate::utils::async_command(git_bin)
+        .args(["apply", "--3way", "--whitespace=nowarn", "-"])
+        .current_dir(&parent_root)
+        .env("PATH", git_env_path())
+        .stdin(Stdio::piped())
+        .stdout(Stdio::piped())
+        .stderr(Stdio::piped())
+        .spawn()
+        .map_err(|e| format!("Failed to run git: {e}"))?;
+
+    if let Some(mut stdin) = child.stdin.take() {
+        stdin
+            .write_all(&patch)
+            .await
+            .map_err(|e| format!("Failed to write git apply input: {e}"))?;
+    }
+
+    let output = child
+        .wait_with_output()
+        .await
+        .map_err(|e| format!("Failed to run git: {e}"))?;
+
+    if output.status.success() {
+        return Ok(());
+    }
+
+    let stderr = String::from_utf8_lossy(&output.stderr);
+    let stdout = String::from_utf8_lossy(&output.stdout);
+    let detail = if stderr.trim().is_empty() {
+        stdout.trim()
+    } else {
+        stderr.trim()
+    };
+    if detail.is_empty() {
+        return Err("Git apply failed.".to_string());
+    }
+
+    if detail.contains("Applied patch to") {
+        if detail.contains("with conflicts") {
+            return Err(
+                "Applied with conflicts. Resolve conflicts in the parent repo before retrying."
+                    .to_string(),
+            );
+        }
+        return Err(
+            "Patch applied partially. Resolve changes in the parent repo before retrying."
+                .to_string(),
+        );
+    }
+
+    Err(detail.to_string())
 }
